chore: extract styles into single files
