+++ frontend/src/component/project/Project/CreateProject/NewCreateProjectForm/ConfigButtons/ChangeRequestTableConfigButton.styles.tsx
@@ -1,0 +1,11 @@
+import { styled } from '@mui/material';
+import { StyledDropdownSearch } from './shared.styles';
+
+export const TableSearchInput = styled(StyledDropdownSearch)({
+    maxWidth: '30ch',
+});
+
+export const ScrollContainer = styled('div')({
+    width: '100%',
+    overflow: 'auto',
+});
