+++ frontend/src/component/project/Project/CreateProject/NewCreateProjectForm/ConfigButtons/ConfigButton.styles.tsx
@@ -1,0 +1,27 @@
+import { Popover, styled } from '@mui/material';
+import { visuallyHiddenStyles } from './shared.styles';
+
+export const StyledDropdown = styled('div')(({ theme }) => ({
+    padding: theme.spacing(2),
+    display: 'flex',
+    flexDirection: 'column',
+    gap: theme.spacing(1),
+    maxHeight: '70vh',
+}));
+
+export const StyledPopover = styled(Popover)(({ theme }) => ({
+    '& .MuiPaper-root': {
+        borderRadius: `${theme.shape.borderRadiusMedium}px`,
+    },
+}));
+
+export const HiddenDescription = styled('p')(() => ({
+    ...visuallyHiddenStyles,
+    position: 'absolute',
+}));
+
+export const ButtonLabel = styled('span', {
+    shouldForwardProp: (prop) => prop !== 'labelWidth',
+})<{ labelWidth?: string }>(({ labelWidth }) => ({
+    width: labelWidth || 'unset',
+}));
